requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
openai==1.12.0
spacy==3.7.2
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
textblob==0.17.1
geopy==2.4.0
# Enhanced dependencies for production
Flask-SocketIO==5.3.6
flasgger==0.9.7.1
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
alembic==1.13.1
marshmallow==3.20.1
webargs==8.3.0
sentry-sdk[flask]==1.38.0
prometheus-client==0.19.0
httpx==0.25.2

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
# ===========================================
# GEN AI LLM DEPENDENCIES
# ===========================================

# Multi-Provider AI APIs
anthropic==0.25.0
google-genai
cohere==4.57
openai==1.3.5

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.25.2

# Image Processing & Computer Vision
Pillow==10.2.0
opencv-python==4.9.0.80
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2

# Audio Processing
openai-whisper
pyttsx3==2.90
gTTS==2.4.0
SpeechRecognition==3.10.1
pyaudio==0.2.11

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2

# Web Scraping & Knowledge
beautifulsoup4==4.12.2
wikipedia==1.4.0
scrapy==2.11.0
newspaper3k==0.2.8

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Machine Learning & Analytics
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Async & Performance
aiohttp==3.9.1
asyncio==3.4.3
uvloop==0.19.0

# Additional Utilities
python-magic==0.4.27
filetype==1.2.0
python-multipart==0.0.6